got table working with values
