+++ bangazon_site/bangazon_storefront/models/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["customer_model","paymenttype_model", 'products_model', 'order_model',  'productonorder_model', 'product_types_model']
+__all__ = ["customer_model","paymenttype_model", 'products_model', 'order_model', 'product_types_model']
