+++ bangazon_site/bangazon_storefront/models/order_model.py
@@ -1,5 +1,5 @@
 from django.db import models
-from . import customer_model, products_model, paymenttype_model
+from . import customer_model, products_model, paymenttype_model, products_model
 
 
 class Order(models.Model):
@@ -13,7 +13,20 @@ class Order(models.Model):
 
     Author: Zoe LeBlanc, Main Bananas
     '''
+    products = models.ManyToManyField('ProductsModel', through='Product_On_Order')
     date_created = models.DateTimeField(auto_now_add=True)
     buyer = models.ForeignKey(customer_model.Customer, null=True, on_delete=models.CASCADE, related_name='orders')
     payment_type = models.ForeignKey(paymenttype_model.PaymentType, blank=True, null=True, on_delete=models.CASCADE)
     payment_complete = models.BooleanField(default=False)
+
+class Product_On_Order(models.Model):
+    ''' 
+    The Product On Order class is a model that defines a join table for Product & Order.
+
+    Argument List:
+        -models.Model: This argument allows the class to access field types.
+
+    Author: Zoe LeBlanc, Main Bananas
+    '''
+    product = models.ForeignKey(products_model.ProductsModel, null=True, related_name='products_on_order')
+    order = models.ForeignKey(Order, null=True, related_name='products_on_order')
