+++ bangazon_site/bangazon_storefront/views/createorder_view.py
@@ -3,7 +3,8 @@ from django.views.generic.base import TemplateView
 from django.http import HttpResponse, HttpResponseRedirect
 from django.contrib.auth.models import User
 from bangazon_storefront.models import *
-
+from bangazon_storefront.models.products_model import *
+from django.db.models import F, Sum, Count
 
 def display_order_and_products(request):
     """
@@ -11,10 +12,9 @@ def display_order_and_products(request):
     """
     customer = customer_model.Customer.objects.get(user=request.user)
     order = order_model.Order.objects.get_or_create(buyer = customer, payment_complete=0)
-    order = order[0]
-    products = productonorder_model.Product_On_Order.objects.filter(order_id=order.pk)
-    order.products = []
-    for product in products:
-        order.products.extend(products_model.ProductsModel.objects.filter(pk=product.pk))
-    context = {'order': order}
+    order_item = order[0]
+    products = order_item.products.all().values().annotate(Count('id')).order_by('id')
+    total =  order_item.products.aggregate(Sum('price'))
+   
+    context = {'order': order_item, 'products': products, 'total': total['price__sum']}
     return render(request, 'bangazon_storefront/order.html', context)
